35강. 애플리케이션 대신 db로 기능 구현하기

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -5,9 +5,8 @@ import com.group.libraryapp.domain.book.BookRepository
 import com.group.libraryapp.domain.book.BookType
 import com.group.libraryapp.domain.user.*
 import com.group.libraryapp.domain.user.loanhistory.UserLoanHistoryRepository
-import com.group.libraryapp.dto.book.BookRequest
-import com.group.libraryapp.dto.book.BookStatResponse
-import com.group.libraryapp.service.book.BookService
+import com.group.libraryapp.dto.book.request.BookRequest
+import com.group.libraryapp.dto.book.response.BookStatResponse
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
@@ -78,7 +77,7 @@ class BookServiceTest @Autowired constructor(
 
     }
 
-    private fun assertCount(result: List<BookStatResponse>, type: BookType, expectedCount: Int) {
+    private fun assertCount(result: List<BookStatResponse>, type: BookType, expectedCount: Long) {
         assertThat(result.first { it.type == type }.count).isEqualTo(expectedCount)
     }
